in BlockMap::for_each_block_under_path, get OS error correctly
not sure if this was changed by WalkDir recently, but
Error::from(<walkdir error>) does not give you the correct underlying
io::Error, but one with more context that wraps it. This IS what we want
to return, but it doesn't have thew right OS error code, so the ENOENT
check here fails, and we emit a scary error message to the log for no
good reason.